pyproject.toml
[tool.poetry]
name = "transcriber"
version = "0.1.0"
description = ""
authors = ["Nicholas Kruckenberg <nkruckenberg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
yt-dlp = "^2023.9.24"
python-dotenv = "^1.0.0"
openai = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
